open-ai-agent-sdk/1.hello-agent/pyproject.toml
[project]
name = "hello-agents"
version = "0.1.0"
description = 'Simple agents that uses the OpenAI Agent SDK to respond to user input.'
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.95.1",
    "openai-agents>=0.1.0",
    "python-dotenv>=1.1.1",
]

authors = [
    { name = "Abdul Rafay", email = "rafaynadeem77@gmail.com" }
]